+++ .gitignore
@@ -52,3 +52,4 @@ libnear_da_rpc_sys.*
 # Ignore configuration for DVN or Executor
 **/config_dvn.toml
 **/config_executor.toml
+**/abi/
