--- workers/src/codec/mod.rs
@@ -1,2 +1,0 @@
-//pub mod bytes_utils;
-//pub mod packet_v1_codec;
