+++ workers/src/data/mod.rs
@@ -1,2 +1,2 @@
 pub mod dvn;
-
+pub mod extractors;
